+++ app/(header)/club/[clubId]/manage/page.tsx
