+++ app/(header)/club/[clubId]/schedule/ScheduleLeagueModal.tsx
