+++ app/(header)/club/[clubId]/schedule/ScheduleOneGameInfo.tsx
