+++ app/(header)/club/[clubId]/schedule/ScheduleOneGameResult.tsx
