+++ app/(header)/club/[clubId]/schedule/[leagueId]/page.tsx
@@ -171,16 +171,16 @@ function LeaguePage() {
         </div>
       </div>
       <div className="flex justify-center items-center pt-8 gap-4">
-        <Link className="w-1/4" href={"/my-club/schedule/1/match"}>
-          <Button
-            size="lg"
-            variant="outline"
-            className="items-center justify-center gap-2 border-primary w-full"
-          >
-            <BookUser size={20} />
-            대진표 보기
-          </Button>
-        </Link>
+        {/* <Link className="w-1/4" href={"/my-club/schedule/1/match"}> */}
+        <Button
+          size="lg"
+          variant="outline"
+          className="items-center justify-center gap-2 border-primary w-full"
+        >
+          <BookUser size={20} />
+          대진표 보기
+        </Button>
+        {/* </Link> */}
         <Button
           size="lg"
           variant={isParticipant ? "destructive" : "default"}
