+++ app/(header)/club/[clubId]/schedule/create/page.tsx
