+++ app/(header)/club/[clubId]/schedule/page.tsx
