--- app/(header)/my-club/layout.tsx
@@ -1,3 +1,0 @@
-import ClubLayout from "@/layouts/ClubLayout";
-
-export default ClubLayout;
